feat(auth): secure demo mode isolation for production
- isDemoModeEnabled() now always returns false in production,
  regardless of ENABLE_DEMO_MODE env var setting
- Added loud console warning if ENABLE_DEMO_MODE=true in production
- /demo-signin page now redirects to /signin when demo mode disabled
- Extracted client-side demo signin logic to separate component

This prevents the demo Credentials provider from ever running in
production, which would otherwise bypass email verification.

Closes #133
